New Changes

+++ 02_trivia_api/backend/flaskr/__init__.py
@@ -8,43 +8,65 @@ from models import setup_db, Question, Category
 
 QUESTIONS_PER_PAGE = 10
 
+##2
+
+def paginate_questions(request, selection):
+	page = request.args.get('page', 1, type=int)
+	start = (page - 1) * QUESTIONS_PER_PAGE
+	end = start + QUESTIONS_PER_PAGE
+	
+	questions = [question.format() for question in selection]
+	current_questions = questions[start:end]
+	
+	return current_questions 
+	
+
+
 def create_app(test_config=None):
-  # create and configure the app
+	# create and configure the app
 	app = Flask(__name__)
 	setup_db(app)
-  
+	app.run(debug= True)
+	
 	'''
+	---1---
 	@TODO: Set up CORS. Allow '*' for origins. Delete the sample route after completing the TODOs
 	'''
 	CORS(app, resources={r"/*": {"origins": "*"}})
 	'''
 	@TODO: Use the after_request decorator to set Access-Control-Allow
 	'''
-    @app.after_request
-    def after_request(response):
-        response.headers.add('Access-Control-Allow-Headers',
-                             'Content-Type,Authorization')
-        response.headers.add('Access-Control-Allow-Methods',
-                             'GET,PUT,POST,DELETE,OPTIONS')
-        return response
+	@app.after_request
+	def after_request(response):
+		response.headers.add('Access-Control-Allow-Headers', 'Content-Type,Authorization')
+		response.headers.add('Access-Control-Allow-Methods', 'GET,PUT,POST,DELETE,OPTIONS')
+		return response
+
 	'''
+	----2----
 	@TODO: 
 	Create an endpoint to handle GET requests 
 	for all available categories.
 	'''
-    @app.route('/categories')
-    def retrieve_categories():
-        categories = Category.query.order_by(Category.type).all()
+	@app.route('/categories')
 
-        if len(categories) == 0:
-            abort(404)
-
-        return jsonify({
-            'success': True,
-            'categories': {category.id: category.type for category in categories}
-        })
-
+	def get_categories():
+		
+		categories = Category.query.all()
+		categories_dict = {}
+		for category in categories:
+			categories_dict[category.id] = category.type
+		
+		if (len(categories_dict) == 0):
+			abort(404)
+	
+		return jsonify({
+			'success': True,
+			'categories': categories_dict
+		})
+		
 	'''
+	----3----
 	@TODO: 
 	Create an endpoint to handle GET requests for questions, 
 	including pagination (every 10 questions). 
@@ -56,46 +78,63 @@ def create_app(test_config=None):
 	ten questions per page and pagination at the bottom of the screen for three pages.
 	Clicking on the page numbers should update the questions. 
 	'''
+	
 	@app.route('/questions')
-    def retrieve_questions():
-        selection = Question.query.order_by(Question.id).all()
-        current_questions = paginate_questions(request, selection)
-
-        categories = Category.query.order_by(Category.type).all()
-
-        if len(current_questions) == 0:
-            abort(404)
-
-        return jsonify({
-            'success': True,
-            'questions': current_questions,
-            'total_questions': len(selection),
-            'categories': {category.id: category.type for category in categories},
-            'current_category': None
-        })
-	
-	
+		
+	def get_questions():
+		'''
+		Handles GET requests for getting all questions.
+		'''
+		
+		questions = Question.query.all()
+		questions_len = len(questions)
+		current_questions = paginate_questions(request, questions)
+		
+		categories = Category.query.all()
+		categories_dict = {}
+		for category in categories:
+			categories_dict[category.id] = category.type
+		
+		if (len(current_questions) == 0):
+			abort(404)
+		
+		return jsonify({
+			'success': True,
+			'questions': current_questions,
+			'total_questions': questions_len,
+			'categories': categories_dict
+		})
+		
+		
 	
 	'''
+	----4----
 	@TODO: 
 	Create an endpoint to DELETE question using a question ID. 
 	
 	TEST: When you click the trash icon next to a question, the question will be removed.
 	This removal will persist in the database and when you refresh the page. 
 	'''
-	@app.route("/questions/<question_id>", methods=['DELETE'])
-    def delete_question(question_id):
-        try:
-            question = Question.query.get(question_id)
-            question.delete()
-            return jsonify({
-                'success': True,
-                'deleted': question_id
-            })
-        except:
-            abort(422)
+	
+	@app.route("/questions/<int:question_id>", methods=['DELETE'])
+	
+	def delete_question(question_id):
+		try:
+			question = Question.query.get(question_id)
+			if question:
+				question.delete()
+			else:
+				absort(404)
+			return jsonify({
+				'success': True,
+				'deleted': question_id
+			})
+		except:
+			abort(422)
+	
 	'''
-	@TODO: 
+	----5----
+	TODO: 
 	Create an endpoint to POST a new question, 
 	which will require the question and answer text, 
 	category, and difficulty score.
@@ -104,84 +143,103 @@ def create_app(test_config=None):
 	the form will clear and the question will appear at the end of the last page
 	of the questions list in the "List" tab.  
 	'''
-	
+	@app.route('/questions', methods=['POST'])
 	def add_question():
-		body = request.get_json()
-
-        if not ('question' in body and 'answer' in body and 'difficulty' in body and 'category' in body):
-            abort(422)
-
-        new_question = body.get('question')
-        new_answer = body.get('answer')
-        new_difficulty = body.get('difficulty')
-        new_category = body.get('category')
-
-        try:
-            question = Question(question=new_question, answer=new_answer,
-                                difficulty=new_difficulty, category=new_category)
-            question.insert()
-
-            return jsonify({
-                'success': True,
-                'created': question.id,
-            })
-
-        except:
-            abort(422)
+		body = request.get_json()		
+		keys = ['question','answer', 'difficulty', 'category' ]
+		if all([key in body.keys() for key in keys]):
+			
+		
+			new_question = body['question']
+			new_answer = body['answer']
+			new_category = body['category']
+			new_difficulty = body['difficulty']
 	
-	'''
-	@TODO: 
-	Create a POST endpoint to get questions based on a search term. 
-	It should return any questions for whom the search term 
-	is a substring of the question. 
+			try:
+				question = Question(question=new_question, answer=new_answer,
+									difficulty=new_difficulty, category=new_category)
+				question.insert()
+		
+				return jsonify({
+					'success': True,
+					'created': question.id,
+				})
 	
-	TEST: Search by any phrase. The questions list will update to include 
-	only question that include that string within their question. 
-	Try using the word "title" to start. 
-	'''
-   @app.route('/questions/search', methods=['POST'])
-    def search_questions():
-        body = request.get_json()
-        search_term = body.get('searchTerm', None)
-
-        if search_term:
-            search_results = Question.query.filter(
-                Question.question.ilike(f'%{search_term}%')).all()
-
-            return jsonify({
-                'success': True,
-                'questions': [question.format() for question in search_results],
-                'total_questions': len(search_results),
-                'current_category': None
-            })
-        abort(404)
+			except:
+				abort(422)
+		else:
+			abort(422)
+	
 
 	
+
 	'''
-	@TODO: 
+	----6----
+	Create a POST endpoint to get questions based on category.
+	'''
+	@app.route('/questions/<int:id>/categorysearch')
+	def get_questions_by_category(id):
+		'''
+		Handles GET requests for getting questions based on category.
+		'''
+		
+		# get the category by id
+		category = Category.query.filter(Category.id == id).all()
+		#return str(category[0].id)
+
+			
+		# get the data if it matches
+		if category:
+		
+			selection = Question.query.filter(Question.category == id).all()	
+			paginated = paginate_questions(request, selection)
+			#return str(paginated[0])
+			return jsonify({
+				'success': True,
+				'questions': paginated,
+				'total_questions': len(Question.query.all()),
+				'current_category': category[0].type
+			})
+		# abort 400 for bad request if category isn't found
+		else:
+			absort(400)
+			
+	'''
+	@TODO:
+	
+    ----7----	
+	
 	Create a GET endpoint to get questions based on category. 
 	
 	TEST: In the "List" tab / main screen, clicking on one of the 
 	categories in the left column will cause only questions of that 
 	category to be shown. 
-	'''
-	@app.route('/categories/<int:category_id>/questions', methods=['GET'])
-    def retrieve_questions_by_category(category_id):
-
-        try:
-            questions = Question.query.filter(
-                Question.category == str(category_id)).all()
-
-            return jsonify({
-                'success': True,
-                'questions': [question.format() for question in questions],
-                'total_questions': len(questions),
-                'current_category': category_id
-            })
-        except:
-            abort(404)
 	
 	'''
+	@app.route('/questions/termsearch', methods=['POST'])
+	def search_questions():
+		body = request.get_json()
+		term = body.get('searchTerm', None)
+
+		if term:
+			search = "%{}%".format(term)			
+			search_results = Question.query.filter(Question.question.like(search)).all()
+			paginated = paginate_questions(request, selection)
+		#	return search_results.questions
+			return jsonify({
+				'success': True,
+				'questions': [question.format() for question in search_results],
+				'total_questions': len(search_results),
+				'current_category': None
+			})
+		else:
+			abort(404)
+
+	
+	
+
+	'''
+	----8----
 	@TODO: 
 	Create a POST endpoint to get questions to play the quiz. 
 	This endpoint should take category and previous question parameters 
@@ -192,64 +250,73 @@ def create_app(test_config=None):
 	one question at a time is displayed, the user is allowed to answer
 	and shown whether they were correct or not. 
 	'''
+	
 	@app.route('/quizzes', methods=['POST'])
-    def play_quiz():
-
-        try:
-
-            body = request.get_json()
-
-            if not ('quiz_category' in body and 'previous_questions' in body):
-                abort(422)
-
-            category = body.get('quiz_category')
-            previous_questions = body.get('previous_questions')
-
-            if category['type'] == 'click':
-                available_questions = Question.query.filter(
-                    Question.id.notin_((previous_questions))).all()
-            else:
-                available_questions = Question.query.filter_by(
-                    category=category['id']).filter(Question.id.notin_((previous_questions))).all()
-
-            new_question = available_questions[random.randrange(
-                0, len(available_questions))].format() if len(available_questions) > 0 else None
-
-            return jsonify({
-                'success': True,
-                'question': new_question
-            })
-        except:
-            abort(422)
+	def play_the_quiz():
+	
+		try:
+	
+			body = request.get_json()
+			keys = ['quiz_category','previous_questions']
+			questions = []
+			if all([key in body.keys() for key in keys]):
+	
+				questions = Question.query.filter(category==category['id']).all()				
+				
+				if len(questions)==0：
+					absort(404)
+					
+				else:
+					questions = [question for question in questions if question.id not in body['previous_questions']]
+					random_selected_question = random.choice(questions).format()
+					
+	
+					return jsonify({
+						'success': True,
+						'question': random_selected_question})
+			else:
+				abort(400)
+		except:
+			abort(422)
+	
 	'''
 	@TODO: 
 	Create error handlers for all expected errors 
 	including 404 and 422. 
 	'''
+
 	@app.errorhandler(404)
-    def not_found(error):
-        return jsonify({
-            "success": False,
-            "error": 404,
-            "message": "resource not found"
-        }), 404
-
-    @app.errorhandler(422)
-    def unprocessable(error):
-        return jsonify({
-            "success": False,
-            "error": 422,
-            "message": "unprocessable"
-        }), 422
-
-    @app.errorhandler(400)
-    def bad_request(error):
-        return jsonify({
-            "success": False,
-            "error": 400,
-            "message": "bad request"
-        }), 400
-  
+	def not_found(error):
+		return jsonify({
+			"success": False,
+			"error": 404,
+			"message": "resource not found"
+		}), 404
+	
+	@app.errorhandler(422)
+	def unprocessable(error):
+		return jsonify({
+			"success": False,
+			"error": 422,
+			"message": "unprocessable"
+		}), 422
+	
+	@app.errorhandler(400)
+	def bad_request(error):
+		return jsonify({
+			"success": False,
+			"error": 400,
+			"message": "bad request"
+		}), 400
+		
+	@app.errorhandler(500)
+	def bad_request(error):
+		return jsonify({
+			"success": False,
+			"error": 500,
+			"message": "Internal Server Error"
+		}), 500
+	
 	return app
-
-    
+	
+
